requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.3
PyMySQL==1.1.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
